test: NewWorksFromGalleriesYouFollow
