+++ src/schema/v2/homeView/__tests__/HomeViewSection.test.ts
@@ -249,93 +249,6 @@ describe("HomeViewSection", () => {
     })
   })
 
-  describe("NewWorksFromGalleriesYouFollow", () => {
-    it("returns lists of artworksConnection", async () => {
-      const query = gql`
-        {
-          homeView {
-            section(
-              id: "home-view-section-new-works-from-galleries-you-follow"
-            ) {
-              __typename
-              component {
-                title
-                behaviors {
-                  viewAll {
-                    buttonText
-                    href
-                    ownerType
-                  }
-                }
-              }
-              ownerType
-
-              ... on HomeViewSectionArtworks {
-                artworksConnection(first: 2) {
-                  edges {
-                    node {
-                      id
-                      title
-                    }
-                  }
-                }
-              }
-            }
-          }
-        }
-      `
-
-      const artworks = [
-        { id: "percy", title: "Percy the Cat" },
-        { id: "matt", title: "Matt the Person" },
-        { id: "paul", title: "Paul the snail" },
-        { id: "paula", title: "Paula the butterfly" },
-      ]
-
-      const context = {
-        accessToken: "424242",
-        followedProfilesArtworksLoader: jest
-          .fn()
-          .mockReturnValue({ body: artworks, headers: { "x-total-count": 2 } }),
-      }
-
-      const { homeView } = await runQuery(query, context)
-
-      expect(homeView.section).toMatchInlineSnapshot(`
-        Object {
-          "__typename": "HomeViewSectionArtworks",
-          "artworksConnection": Object {
-            "edges": Array [
-              Object {
-                "node": Object {
-                  "id": "QXJ0d29yazpwZXJjeQ==",
-                  "title": "Percy the Cat",
-                },
-              },
-              Object {
-                "node": Object {
-                  "id": "QXJ0d29yazptYXR0",
-                  "title": "Matt the Person",
-                },
-              },
-            ],
-          },
-          "component": Object {
-            "behaviors": Object {
-              "viewAll": Object {
-                "buttonText": "Browse All Artworks",
-                "href": null,
-                "ownerType": null,
-              },
-            },
-            "title": "New Works from Galleries You Follow",
-          },
-          "ownerType": "newWorksFromGalleriesYouFollow",
-        }
-      `)
-    })
-  })
-
   describe("RecommendedArtists", () => {
     it("returns lists of artists with the right title", async () => {
       const query = gql`
